# [feat] Add a lightweight time profiler in the framework