@vkarak 2022-06-30
> Let's imagine that a user would like to also measure their own test hooks...
The timings would be per test, not per test case, right? If that's the case, how could someone add a test case-specific timing?

The timings depend on the user. If you use the `@time_function` decorator, then yes, it will per class instance and it will show the cumulative hook time for all object instantiated from this class, i.e., for every test case. If you want to have separate timing per case, you should use the `time_region` context manager and name the region either after the unique or the display name of the test.

> The intention is to measure the framework, right? That's the reason why the _profiler is defined in the core/logging.py file... However, would it make sense to have a timing function on the utilities? Especially to help with question 1?

Yes, the intention is to measure the framework performance and not to make available –at least for now– the profiler to the users. Let's concrete cases where this could be useful and we can "open it up."